# openapi.yaml
openapi: 3.0.3
info:
  title: Club Operations POS API
  version: 1.0.0
  description: |
    Server-authoritative REST API for managing club check-ins, room/locker assignments,
    cleaning workflows, checkout verification, and operational reporting.

    All endpoints are versioned under `/v1`.
    Request bodies are validated with Zod schemas.

    Authentication:
    - Staff endpoints require Bearer token (session token from PIN or WebAuthn login)
    - Admin endpoints require ADMIN role in addition to authentication
    - Customer-facing endpoints (agreement signing, checkout kiosk) do not require auth
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /v1/auth/login-pin:
    post:
      tags: [Authentication]
      summary: Staff login with PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [staffLookup, deviceId, pin]
              properties:
                staffLookup:
                  type: string
                  description: Staff ID or name
                deviceId:
                  type: string
                pin:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  staffId: { type: string, format: uuid }
                  name: { type: string }
                  role: { type: string, enum: [STAFF, ADMIN] }
                  sessionToken: { type: string }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  message: { type: string }
...
  /v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Staff logout (revoke current session token)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  message: { type: string }
  /v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get current staff identity
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current staff identity
          content:
            application/json:
              schema:
                type: object
                properties:
                  staffId: { type: string, format: uuid }
                  name: { type: string }
                  role: { type: string, enum: [STAFF, ADMIN] }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  message: { type: string }
  /v1/auth/reauth-pin:
    post:
      tags: [Authentication]
      summary: Re-authenticate with PIN for sensitive actions (sets reauth_ok_until)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin:
                  type: string
      responses:
        '200':
          description: Re-auth successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  reauthOkUntil: { type: string, format: date-time }
        '400':
          description: Validation failed
        '401':
          description: Unauthorized
  /v1/auth/reauth/webauthn/options:
    post:
      tags: [Authentication]
      summary: Get WebAuthn re-authentication options
      security:
        - bearerAuth: []
      responses:
        '200':
          description: WebAuthn authentication options (server-generated)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: No passkeys registered / invalid request
        '401':
          description: Unauthorized
  /v1/auth/reauth/webauthn/verify:
    post:
      tags: [Authentication]
      summary: Verify WebAuthn re-authentication (sets reauth_ok_until)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credentialResponse]
              properties:
                credentialResponse:
                  type: object
                  additionalProperties: true
                deviceId:
                  type: string
      responses:
        '200':
          description: Re-auth successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  reauthOkUntil: { type: string, format: date-time }
        '400':
          description: Invalid/expired challenge or verification failed
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque session token from PIN or WebAuthn login (stored/validated server-side; not a JWT)
...